requirements.txt
rich
bcrypt
customtkinter
Pillow
CTkTable
